runtime: fix two latent correctness bugs found in code review
JUMP32 / JUMP32_FALSE / JUMP32_TRUE (interpreter) assembled their 32-bit
offset with `(hi << 16) | lo` and passed it through `as isize`, which
ZERO-extends the u32 — so backward long jumps (negative offsets) jumped
forward by ~4 GB instead. Upstream bytecode.cpp:2236 sign-extends. Cast
through `i32` before widening. Only fires for backward 32-bit jumps inside
functions larger than 64 KB, which no current workload contains — the
1300-case differential oracle never exercised it.

GC: an untracked from-space pointer (a slot the collector failed to
forward) was logged and then ignored — `replace_storage` immediately drops
from-space, leaving that slot dangling (silent use-after-free). In correct
operation this set is always empty, so make a non-empty set a hard panic:
crash loudly on a collector bug rather than corrupt the heap.

+++ crates/polyml-runtime/src/interpreter/mod.rs
@@ -2906,27 +2906,31 @@ impl Interpreter {
 
             // ----- Wider jumps / case
             EXTINSTR_JUMP32 => {
+                // The 32-bit offset is SIGNED — upstream sign-extends it
+                // (bytecode.cpp:2236). `as i32 as isize` sign-extends so a
+                // backward long jump stays negative (was zero-extended → ~4GB
+                // forward jump on backward jumps in functions >64KB).
                 let lo = u32::from(self.fetch_u16_le()?);
                 let hi = u32::from(self.fetch_u16_le()?);
-                let off = (hi << 16) | lo;
-                self.pc_offset_signed(off as isize)?;
+                let off = ((hi << 16) | lo) as i32 as isize;
+                self.pc_offset_signed(off)?;
                 Ok(StepResult::Continue)
             }
             EXTINSTR_JUMP32_FALSE => {
                 let lo = u32::from(self.fetch_u16_le()?);
                 let hi = u32::from(self.fetch_u16_le()?);
-                let off = (hi << 16) | lo;
+                let off = ((hi << 16) | lo) as i32 as isize; // signed (see JUMP32)
                 if self.pop()? == PolyWord::tagged(0) {
-                    self.pc_offset_signed(off as isize)?;
+                    self.pc_offset_signed(off)?;
                 }
                 Ok(StepResult::Continue)
             }
             EXTINSTR_JUMP32_TRUE => {
                 let lo = u32::from(self.fetch_u16_le()?);
                 let hi = u32::from(self.fetch_u16_le()?);
-                let off = (hi << 16) | lo;
+                let off = ((hi << 16) | lo) as i32 as isize; // signed (see JUMP32)
                 if self.pop()? != PolyWord::tagged(0) {
-                    self.pc_offset_signed(off as isize)?;
+                    self.pc_offset_signed(off)?;
                 }
                 Ok(StepResult::Continue)
             }
